requirements.txt
Flask
Flask-Login
dnspython
email-validator
